Golden_Chain/pinkblonsonbeta/requirements.txt
google-generativeai
Pillow
psutil
python-dotenv
sounddevice
soundfile
numpy
faster-whisper
